add projectinputs

+++ stocks-api/src/main/kotlin/com/cleverskills/api/ApiCreateProjectRequest.kt
@@ -18,7 +18,7 @@ data class ApiProjectInputs(
     @JsonProperty("prix") val prix: Long,
     @JsonProperty("travaux") val travaux: Long,
     @JsonProperty("apport") val apport: Long,
-    @JsonProperty("tauxCredit") val tauxCredit: Long,
+    @JsonProperty("tauxCredit") val tauxCredit: Double,
     @JsonProperty("dureeCredit") val dureeCredit: Long,
     @JsonProperty("meubles") val meubles: Long,
     @JsonProperty("copro") val copro: Long,
